+++ Parley/CHANGELOG.md
@@ -10,6 +10,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.1.56-alpha] - 2025-12-12
+**Branch**: `parley/sprint/techdebt` | **PR**: #TBD
+
+### Sprint: Parley Techdebt (#347)
+
+- TBD
+
+---
+
 ## [0.1.55-alpha] - 2025-12-12
 **Branch**: `parley/refactor/safe-control-finder` | **PR**: #348